Login and maintenance enhancements

+++ src/maintenance.psm1
@@ -408,7 +408,7 @@ function Invoke-SafeguardApplianceShutdown
 
     if ($Force)
     {
-        $Confirmed = $true
+        $local:Confirmed = $true
     }
     else
     {
@@ -509,14 +509,18 @@ function Invoke-SafeguardApplianceReboot
 
     if ($local:Confirmed)
     {
+        if (-not $NoWait)
+        {
+            Import-Module -Name "$PSScriptRoot\sg-utilities.psm1" -Scope Local
+            $local:CurrentState = (Get-SafeguardStatus -Appliance $Appliance -Insecure:$Insecure).ApplianceCurrentState
+        }
         Write-Host "Sending reboot command..."
         Invoke-SafeguardMethod -AccessToken $AccessToken -Appliance $Appliance -Insecure:$Insecure Appliance `
             POST ApplianceStatus/Reboot -Body $Reason
 
         if (-not $NoWait)
         {
-            Import-Module -Name "$PSScriptRoot\sg-utilities.psm1" -Scope Local
-            Wait-ForSafeguardOnlineStatus -Appliance $Appliance -Insecure:$Insecure -Timeout $Timeout
+            Wait-ForSafeguardStatus -Appliance $Appliance -Insecure:$Insecure -Timeout $Timeout -DesiredStatus $local:CurrentState
         }
     }
     else
@@ -583,7 +587,7 @@ function Invoke-SafeguardApplianceFactoryReset
 
     if ($Force)
     {
-        $Confirmed = $true
+        $local:Confirmed = $true
     }
     else
     {
@@ -901,6 +905,9 @@ Use the currently staged patch rather than uploading a new one.
 .PARAMETER NoWait
 Specify this flag to continue immediately without waiting for the patch to install to the connected appliance.
 
+.PARAMETER Force
+Do not prompt for confirmation.
+
 .INPUTS
 None.
 
@@ -929,7 +936,9 @@ function Install-SafeguardPatch
         [Parameter(ParameterSetName="UseExisting",Mandatory=$false)]
         [switch]$UseStagedPatch = $false,
         [Parameter(Mandatory=$false)]
-        [switch]$NoWait
+        [switch]$NoWait,
+        [Parameter(Mandatory=$false)]
+        [switch]$Force
     )
 
     $ErrorActionPreference = "Stop"
@@ -1028,11 +1037,18 @@ function Install-SafeguardPatch
         Wait-ForPatchDistribution -AccessToken $AccessToken -Appliance $Appliance -Insecure:$Insecure
     }
 
-    Import-Module -Name "$PSScriptRoot\ps-utilities.psm1" -Scope Local
-    $local:Confirmed = (Get-Confirmation "Install Safeguard Patch" `
-                                         "Do you want to install $($local:StagedPatch.Title) on this cluster?" `
-                                         "Starts cluster patch immediately." `
-                                         "Cancels this operation.")
+    if ($Force)
+    {
+        $local:Confirmed = $true
+    }
+    else
+    {
+        Import-Module -Name "$PSScriptRoot\ps-utilities.psm1" -Scope Local
+        $local:Confirmed = (Get-Confirmation "Install Safeguard Patch" `
+                                            "Do you want to install $($local:StagedPatch.Title) on this cluster?" `
+                                            "Starts cluster patch immediately." `
+                                            "Cancels this operation.")
+    }
     if ($local:Confirmed)
     {
         Write-Host "Starting patch install..."
@@ -1164,8 +1180,8 @@ function Remove-SafeguardBackup
 
     if (-not $BackupId)
     {
-        $CurrentBackupIds = (Get-SafeguardBackup -AccessToken $AccessToken -Appliance $Appliance -Insecure:$Insecure).Id -join ", "
-        Write-Host "Available Backups: [ $CurrentBackupIds ]"
+        $local:CurrentBackupIds = (Get-SafeguardBackup -AccessToken $AccessToken -Appliance $Appliance -Insecure:$Insecure).Id -join ", "
+        Write-Host "Available Backups: [ $($local:CurrentBackupIds) ]"
         $BackupId = (Read-Host "BackupId")
     }
 
@@ -1257,8 +1273,8 @@ function Export-SafeguardBackup
 
     if (-not $BackupId)
     {
-        $CurrentBackupIds = (Get-SafeguardBackup -AccessToken $AccessToken -Appliance $Appliance -Insecure:$Insecure).Id -join ", "
-        Write-Host "Available Backups: [ $CurrentBackupIds ]"
+        $local:CurrentBackupIds = (Get-SafeguardBackup -AccessToken $AccessToken -Appliance $Appliance -Insecure:$Insecure).Id -join ", "
+        Write-Host "Available Backups: [ $($local:CurrentBackupIds) ]"
         $BackupId = (Read-Host "BackupId")
     }
     if (-not $OutFile)
